bug fixes and trello api app routes added

+++ server/trello_backend/trello_api/views.py
@@ -127,46 +127,29 @@ def create_card(request):
         return Response({"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 
-@api_view(["GET"])
-def get_card(request, card_id):
+@api_view(["GET", "PUT", "DELETE"])
+def card_retrieve_update_delete(request, card_id):
     try:
         card = Card.objects.get(pk=card_id)
-        serializer = CardSerializer(card)
-        return Response(serializer.data, status=status.HTTP_200_OK)
-    except Card.DoesNotExist:
-        return Response({"error": "Card not found"}, status=status.HTTP_404_NOT_FOUND)
-    except Exception as e:
-        return Response({"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
-
-@api_view(["PUT"])
-def update_card(request, card_id):
-    try:
-        card = Card.objects.get(pk=card_id)
-        serializer = CardSerializer(instance=card, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
+        if request.method == "GET":
+            serializer = CardSerializer(card)
             return Response(serializer.data, status=status.HTTP_200_OK)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+        elif request.method == "PUT":
+            serializer = CardSerializer(instance=card, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+        elif request.method == "DELETE":
+            card.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
     except Card.DoesNotExist:
         return Response({"error": "Card not found"}, status=status.HTTP_404_NOT_FOUND)
     except Exception as e:
         return Response({"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-
-@api_view(["DELETE"])
-def delete_card(request, card_id):
-    try:
-        card = Card.objects.get(pk=card_id)
-        card.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-    except Card.DoesNotExist:
-        return Response({"error": "Card not found"}, status=status.HTTP_404_NOT_FOUND)
-    except Exception as e:
-        return Response({"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-
-# views.py
 
 
 @api_view(["PUT"])
